Refactor equipo anita

+++ src/Clientes/ListadoCliente.jsx
@@ -1,61 +1,67 @@
-import axios from "axios";
 import React, { useContext, useEffect, useState } from "react";
-import { Link, useParams } from "react-router-dom";
-import { ITEMS_PER_PAGE, API_URL } from "../App.config";
-import { ClienteContext } from "./ClienteContext";
-import { obtenerClientes, eliminarCliente } from "../Services/ClienteService";
+import { Link, useNavigate } from "react-router-dom";
+import { IMAGEN_EDIT, IMAGEN_DELETE, ITEMS_PER_PAGE } from "../App.config";
+import { ClienteContext } from "../Context/ClienteContext";
+import {
+  obtenerClientesPorPagina,
+  eliminarCliente,
+} from "../Services/ClienteService";
 
 export default function ListadoCliente() {
+  let navegacion = useNavigate();
   const { clientes, setClientes } = useContext(ClienteContext);
+
   const [consulta, setConsulta] = useState("");
-  const [page, setPage] = useState(0);
-  const [pageSize, setPageSize] = useState(ITEMS_PER_PAGE);
-  const [totalPages, setTotalPages] = useState(0);
+
+  const [pagina, setPagina] = useState(0);
+  const [tamañoPagina, setTamañoPagina] = useState(ITEMS_PER_PAGE);
+  const [cantidadPaginas, setCantidadPaginas] = useState(0);
+
   const [sortConfig, setSortConfig] = useState({
     key: null,
     direction: "ascending",
-  }); //se utiliza para el orden
+  }); // Se utiliza para el orden
 
   useEffect(() => {
-    getDatos();
-  }, [page, pageSize, consulta]);
+    cargarClientes();
+  }, [pagina, tamañoPagina, consulta]);
 
-  const handlePageChange = (newPage) => {
-    setPage(newPage);
+  const cambiarPagina = (newPage) => {
+    setPagina(newPage);
   };
 
-  const getDatos = async () => {
-    console.log("carga " + page);
-    obtenerClientes(consulta, page, pageSize)
+  const cargarClientes = async () => {
+    //console.log("carga " + pagina);
+    obtenerClientesPorPagina(consulta, pagina, tamañoPagina)
       .then((response) => {
         setClientes(response.content);
-        setTotalPages(response.totalPages);
+        setCantidadPaginas(response.totalPages);
       })
       .catch((error) => {
-        console.error("Error fetching items:", error);
+        console.error("Error al obtener clientes", error);
       });
   };
 
-  const handConsultaChange = (e) => {
+  const cambiarConsulta = (e) => {
     setConsulta(e.target.value);
   };
 
+  const editar = async (id) => {
+    if (window.confirm("¿Estas seguro de que deseas editar este cliente?")) {
+      navegacion(`/cliente/${id}`);
+    }
+  };
+
   const eliminar = async (id) => {
-    try {
-      const eliminacionExitosa = await eliminarCliente(id);
-      if (eliminacionExitosa) {
-        getDatos();
-      } else {
-        console.error("Error al eliminar el cliente");
-      }
-    } catch (error) {
-      console.error("Error al eliminar el cliente:", error);
+    if (window.confirm("¿Estás seguro de que deseas eliminar este cliente?")) {
+      await eliminarCliente(id);
+      cargarClientes();
     }
   };
 
   ///////////////////////////////////////Para el orden de las tablas///////////////////////////////////////////////////
 
-  const handleSort = (key) => {
+  const filtrarClientes = (key) => {
     let direction = "ascending";
     if (sortConfig.key === key && sortConfig.direction === "ascending") {
       direction = "descending";
@@ -63,7 +69,7 @@ export default function ListadoCliente() {
     setSortConfig({ key, direction });
   };
 
-  const sortedData = () => {
+  const clientesFiltrados = () => {
     const sorted = [...clientes];
     if (sortConfig.key !== null) {
       sorted.sort((a, b) => {
@@ -96,12 +102,12 @@ export default function ListadoCliente() {
             type="search"
             aria-label="Search"
             value={consulta}
-            onChange={handConsultaChange}
+            onChange={cambiarConsulta}
           ></input>
         </div>
         <div className="col-1">
           <button
-            onClick={() => getDatos()}
+            onClick={() => cargarClientes()}
             className="btn btn-outline-success"
             type="submit"
           >
@@ -111,9 +117,9 @@ export default function ListadoCliente() {
       </div>
       <hr></hr>
       <table className="table table-striped table-hover align-middle">
-        <thead className="table-dark">
+        <thead className="table-dark text-center">
           <tr>
-            <th scope="col" onClick={() => handleSort("id")}>
+            <th scope="col" onClick={() => filtrarClientes("id")}>
               #
               {sortConfig.key === "id" && (
                 <span>
@@ -121,7 +127,7 @@ export default function ListadoCliente() {
                 </span>
               )}
             </th>
-            <th scope="col" onClick={() => handleSort("razonSocial")}>
+            <th scope="col" onClick={() => filtrarClientes("razonSocial")}>
               Apellido y Nombre
               {sortConfig.key === "razonSocial" && (
                 <span>
@@ -129,7 +135,7 @@ export default function ListadoCliente() {
                 </span>
               )}
             </th>
-            <th scope="col" onClick={() => handleSort("celular")}>
+            <th scope="col" onClick={() => filtrarClientes("celular")}>
               Cel
               {sortConfig.key === "celular" && (
                 <span>
@@ -137,7 +143,7 @@ export default function ListadoCliente() {
                 </span>
               )}
             </th>
-            <th scope="col" onClick={() => handleSort("mail")}>
+            <th scope="col" onClick={() => filtrarClientes("mail")}>
               Mail
               {sortConfig.key === "mail" && (
                 <span>
@@ -148,10 +154,10 @@ export default function ListadoCliente() {
             <th scope="col">Acciones</th>
           </tr>
         </thead>
-        <tbody>
+        <tbody className="text-center">
           {
-            //iteramos empleados
-            sortedData().map((cliente, indice) => (
+            // Iteramos clientesFIltrados para mostrarlos en la tabla
+            clientesFiltrados().map((cliente, indice) => (
               <tr key={indice}>
                 <th scope="row">{cliente.id}</th>
                 <td>{cliente.razonSocial}</td>
@@ -160,18 +166,26 @@ export default function ListadoCliente() {
 
                 <td className="text-center">
                   <div>
-                    <Link
-                      to={`/cliente/${cliente.id}`}
+                    <button
+                      onClick={() => editar(cliente.id)}
                       className="btn btn-link btn-sm me-3"
                     >
+                      <img
+                        src={IMAGEN_EDIT}
+                        style={{ width: "20px", height: "20px" }}
+                      />
                       Editar
-                    </Link>
+                    </button>
 
                     <button
                       onClick={() => eliminar(cliente.id)}
                       className="btn btn-link btn-sm me-3"
                     >
                       {" "}
+                      <img
+                        src={IMAGEN_DELETE}
+                        style={{ width: "20px", height: "20px" }}
+                      />
                       Eliminar
                     </button>
                   </div>
@@ -198,18 +212,20 @@ export default function ListadoCliente() {
       {/* /////////////////////// Esto se utiliza para hacer la paginacion  ///////////////////////////////// */}
 
       <div className="pagination d-md-flex justify-content-md-end">
-        {Array.from({ length: totalPages }, (_, i) => i).map((pageNumber) => (
-          <a
-            key={pageNumber}
-            href="#"
-            onClick={(e) => {
-              e.preventDefault(); // Previene el comportamiento predeterminado del enlace
-              handlePageChange(pageNumber);
-            }}
-          >
-            | {pageNumber} |
-          </a>
-        ))}
+        {Array.from({ length: cantidadPaginas }, (_, i) => i).map(
+          (pageNumber) => (
+            <a
+              key={pageNumber}
+              href="#"
+              onClick={(e) => {
+                e.preventDefault(); // Previene el comportamiento predeterminado del enlace
+                cambiarPagina(pageNumber);
+              }}
+            >
+              | {pageNumber} |
+            </a>
+          )
+        )}
       </div>
 
       {/* /////////////////////// fin de la paginacion  ///////////////////////////////// */}
